+++ src/site/craft_logic.h
@@ -22,7 +22,7 @@ std::vector<flecs::entity> collect_nearby_source_containers(
     SiteRunState& site_run,
     TileCoord device_tile);
 
-std::vector<std::uint32_t> collect_nearby_item_instance_ids(
+std::vector<std::uint64_t> collect_nearby_item_instance_ids(
     SiteRunState& site_run,
     TileCoord device_tile);
 
@@ -34,7 +34,7 @@ const CraftDeviceCacheState* find_device_cache(
     const CraftState& craft_state,
     std::uint64_t device_entity_id) noexcept;
 
-std::vector<std::uint32_t> nearby_item_instance_ids_for_device(
+std::vector<std::uint64_t> nearby_item_instance_ids_for_device(
     SiteRunState& site_run,
     const CraftState& craft_state,
     std::uint64_t device_entity_id,
@@ -42,19 +42,19 @@ std::vector<std::uint32_t> nearby_item_instance_ids_for_device(
 
 std::uint32_t available_cached_item_quantity(
     SiteRunState& site_run,
-    const std::vector<std::uint32_t>& item_instance_ids,
+    const std::vector<std::uint64_t>& item_instance_ids,
     ItemId item_id) noexcept;
 
 bool recipe_requires_hammer(const CraftRecipeDef& recipe_def) noexcept;
 
 bool can_satisfy_recipe_ingredients(
     SiteRunState& site_run,
-    const std::vector<std::uint32_t>& item_instance_ids,
+    const std::vector<std::uint64_t>& item_instance_ids,
     const CraftRecipeDef& recipe_def) noexcept;
 
 bool can_satisfy_recipe_requirements(
     SiteRunState& site_run,
-    const std::vector<std::uint32_t>& item_instance_ids,
+    const std::vector<std::uint64_t>& item_instance_ids,
     const CraftRecipeDef& recipe_def) noexcept;
 
 bool can_store_output_after_recipe_consumption(
